Fix progress bar and spinner
