+++ packages/react/src/components/shared.tsx
@@ -1,7 +1,7 @@
 export function Spinner() {
   return (
     <svg
-      className="block h-5 w-5 animate-spin align-middle text-white"
+      className="z-10 block h-5 w-5 animate-spin align-middle text-white"
       xmlns="http://www.w3.org/2000/svg"
       fill="none"
       viewBox="0 0 576 512"
@@ -13,3 +13,17 @@ export function Spinner() {
     </svg>
   );
 }
+
+export const progressWidths: Record<number, string> = {
+  0: "after:w-0",
+  10: "after:w-[10%]",
+  20: "after:w-[20%]",
+  30: "after:w-[30%]",
+  40: "after:w-[40%]",
+  50: "after:w-[50%]",
+  60: "after:w-[60%]",
+  70: "after:w-[70%]",
+  80: "after:w-[80%]",
+  90: "after:w-[90%]",
+  100: "after:w-[100%]",
+};
